improved full_uber_matrix target
started using a regression to calculate the cost based on the travel
time and the distance between two points. also, we're now prioritising
the values found in the uber data rather than the ones calculated by
dodgr. and finally, whenever dodgr calculates the travel time/distance
from a hex to itself as 0, and the uber data doesn't contain the trips
from that hexagon to itself, we substitute these 0 by the average travel
time and distance of all trips from hexagons to themselves